+++ ui/packages/app/grafana-panel-plugin/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.153 (2023-07-06)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
 ## [0.0.152](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.151...@parca/grafana-panel@0.0.152) (2023-07-05)
 
 **Note:** Version bump only for package @parca/grafana-panel